Add Logo for Croakus (CROAK) Token
This pull request adds the official logo for Croakus (CROAK) token to the Trust Wallet forked repository. The logo file has been uploaded in accordance with the guidelines and specifications outlined by Trust Wallet.

The logo accurately represents the identity of the Croakus (CROAK) token and will enhance the user experience for Trust Wallet users who hold or transact with CROAK.

Thank you for considering this pull request.